feat: Refactor ACR tests (PSKD-1400)
DCO Remediation Commit for nathan.dickens@sas.com <Nathan.Dickens@sas.com>

I, nathan.dickens@sas.com <Nathan.Dickens@sas.com>, hereby add my Signed-off-by to this commit: 0b2e1ceb8e68be705da523fc35e0d7a3a83e7a57
